requirements.txt
alabaster==0.7.10
astroid==1.6.4
Babel==2.5.3
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==2.0.5
django-admin==1.2.4
django-bootstrap-form==3.4
django-bootstrap-forms==0.1
django-bower==5.2.0
django-compat==1.0.15
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-excel-response2==2.0.8
django-grappelli==2.11.1
django-hijack==2.1.7
django-hijack-admin==2.1.7
django-nose==1.4.5
django-six==1.0.4
djangorestframework==3.8.2
docutils==0.14
future==0.16.0
googlemaps==2.5.1
httplib2==0.11.3
httpretty==0.9.4
idna==2.6
imagesize==1.0.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
logilab-common==1.4.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==2.0.7
packaging==17.1
pbr==4.0.3
Pygments==2.2.0
PyJWT==1.6.3
pylint==1.9.1
pyparsing==2.2.0
PySocks==1.6.8
python-openid==2.2.5
PyTumblr==0.0.7
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
screen==1.0.1
simplejson==3.15.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinx-rtd-theme==0.3.1
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
twilio==6.14.0
urllib3==1.22
wrapt==1.10.11
xlwt==1.3.0
